+++ packages/valorant.ts/src/index.ts
@@ -1,3 +1,0 @@
-export { ValCrosshair } from "./utils/ValCrosshair";
-export { ValPatchNote } from "./utils/ValPatchNote";
-export { ValVersion } from "./utils/ValVersion";